requirements.txt
attrs==25.3.0
blinker==1.9.0
cachetools==5.5.2
certifi==2025.8.3
cffi==2.0.0
charset-normalizer==3.4.3
click==8.2.1
colorama==0.4.6
cryptography==41.0.7
Flask==2.3.3
google-ai-generativelanguage==0.4.0
google-api-core==2.25.1
google-auth==2.40.3
google-generativeai==0.3.2
googleapis-common-protos==1.70.0
grpcio==1.75.0
grpcio-status==1.62.3
gunicorn==21.2.0
h11==0.16.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
loguru==0.7.2
MarkupSafe==3.0.2
mercadopago==2.2.0
outcome==1.3.0.post0
packaging==25.0
pillow==11.3.0
proto-plus==1.26.1
protobuf==4.25.8
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==1.0.0
requests==2.31.0
rsa==4.9.1
selenium==4.15.0
setuptools==80.9.0
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
supervisor==4.2.5
tqdm==4.66.1
trio==0.31.0
trio-websocket==0.12.2
typing_extensions==4.15.0
urllib3==1.26.18
webdriver-manager==4.0.1
Werkzeug==2.3.7
win32_setctime==1.2.0
wsproto==1.2.0
